requirements.txt
streamlit==1.38.0
pandas==2.2.2
gspread==5.12.4
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
